+++ History.md
@@ -1,4 +1,4 @@
-### vNext ###
+### 1.5.0 ###
 - Added suppression for member declarations in XAML files (fixes issue #6)
 - Added new setting to allow to ignore overriding type members (implements #3)